Made Health Integers

+++ Assets/Scripts/Enemies/SteamKing/SteamKing.cs
@@ -304,7 +304,7 @@ public class SteamKing : EnemyBase
         }
     }
 
-    public override void TakeDamage(float damage)
+    public override void TakeDamage(int damage)
     {
         if (EnemyManager.instance.LogEnemyDamage)
         {
